+++ lib/contentful/base_resource.rb
@@ -98,7 +98,7 @@ module Contentful
     def hydrate_metadata
       result = {}
       raw.fetch('metadata', {}).each do |k, v|
-        v = v.map { |tag| build_link(tag) } if k == 'tags'
+        v = v.map { |item| build_link(item) } if %w[tags concepts].include?(k)
         result[Support.snakify(k, @configuration[:use_camel_case]).to_sym] = v
       end
       result